requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bitarray==3.0.0
ccxt==4.4.50
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cryptography==44.0.0
cytoolz==1.0.1
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
multidict==6.1.0
numpy==2.2.2
pandas==2.2.3
parsimonious==0.10.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
pywin32==308
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
setuptools==75.8.0
six==1.17.0
tabulate==0.9.0
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
web3==7.7.0
websockets==13.1
yarl==1.18.3
